Update BaseController.cs
I have been exploring your code and it looks really great. I just want to propose this change due that you are using here an API controller, and it would be better to use the ControllerBase class rather than Controller, this is due that ControllerBase doesn't handle thing like response with views. Hope this helps.